Move eds dialog to select modify or create survey to VariantSpecificUtils
